strategy/requirements.txt
pandas==2.1.4
numpy==1.26.3
backtrader==1.9.78.123
vectorbt==0.26.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
flask==3.0.0
kafka-python==2.0.2
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0

